Duplicated data fix & remove point by id.
Duplicated data fix & remove point by id.

+++ site/hBars.php
@@ -14,6 +14,10 @@ if(isset($_REQUEST['get'])) {
 	echo getHBars(@$_REQUEST['latitude'], @$_REQUEST['longitude']);
 }
 
+if(isset($_REQUEST['del'])) {
+	echo removeHBar(@$_REQUEST['id']);
+}
+
 //Добавление турников
 function addHBar($adminpass, $latitude, $longitude, $description, $imgs = '') {
 	global $mysqli, $config;
@@ -24,9 +28,14 @@ function addHBar($adminpass, $latitude, $longitude, $description, $imgs = '') {
 		$longitude = mysqli_real_escape_string($mysqli, $longitude);
 		$description = mysqli_real_escape_string($mysqli, $description);
 		$imgs = mysqli_real_escape_string($mysqli, $imgs);
-		$q = "INSERT INTO `".$config['sqlprefix']."hBars`(`latitude`, `longitude`, `description`, `imgs`) VALUES (".$latitude.",".$longitude.",'".$description."','".$imgs."')";
-		mysqli_query($mysqli, $q)or DIE('OH NO! I did not want to kill him : '.mysqli_error($mysqli));
-		return 'Success!';
+		$q = "SELECT * FROM `".$config['sqlprefix']."hBars` WHERE `latitude` LIKE ".$latitude." AND `longitude` LIKE ".$longitude;
+		if(mysqli_num_rows(mysqli_query($mysqli, $q)) == 0) {
+			$q = "INSERT INTO `".$config['sqlprefix']."hBars`(`latitude`, `longitude`, `description`, `imgs`) VALUES (".$latitude.",".$longitude.",'".$description."','".$imgs."')";
+			mysqli_query($mysqli, $q)or DIE('OH NO! I did not want to kill him : '.mysqli_error($mysqli));
+			return 'Success!';
+		} else {
+			return 'Data is duplicated!';
+		}
 	}
 
 }
@@ -46,5 +55,13 @@ function getHBars($latitude = 0, $longitude = 0) {
 		}
 		return json_encode($result);
 	}
+}
+
+function removeHBar($id) {
+	global $mysqli, $config;
+	$id = mysqli_real_escape_string($mysqli, $id);
+	$q = "DELETE FROM `".$config['sqlprefix']."hBars` WHERE `id` = ".$id;
+	mysqli_query($mysqli, $q)or DIE('OH NO! I did not want to kill him : '.mysqli_error($mysqli));
+	return 'Success!';
 }
 ?>
